improv: otimizado o sistema para receber valores float do wifi

+++ lib/wifi/wifi.hpp
@@ -10,14 +10,14 @@ namespace Wifi{
 
     struct rcv_message{
         int8_t id;
-        int16_t v;
-        int16_t w;
+        float v;
+        float w;
         int32_t checksum;
     };
   
     void setup(uint8_t robot);
     void OnDataRecv(const uint8_t *mac, const uint8_t *incomingData, int len);
-    void receiveData(int16_t *v, int16_t *w);
+    void receiveData(float *v, float *w);
     bool isCommunicationLost();
     void tokenize(const uint8_t *data,int len);
 }
